requirements.txt
# base
django==5.0.3
djangorestframework=3.15.1

# snippets
pygments==2.17.2 # code highlighting